fake_news_detector/requirements.txt
Flask
numpy
Pillow
opencv-python
tensorflow
joblib
scikit-learn

